Procedure: Move test for internal procedures to dedicated test
